[COMMIT] Refactoring layout and codebase
Refactored the structure of the UI code and left it in script but moved all functional code into a new class and script
