+++ Assets/Editor/SOCWizard/SOCWizardLogic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: adb683fdebe8ca3419d2da8b1c3ace07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
