--- Assets/Scenes/test2.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 70241b0ca04eb734fa24740655672862
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
