--- Assets/Scenes/test2.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 429d0c223c0c1b5488d1efc22cbe27f7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
