--- Assets/Scenes/test2/OcclusionCullingData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5b8f4a4b54440b04d8928d4c57c3ed7d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 36300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
