--- Assets/Scenes/test3.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff3d29b63b9baae4f963ef909c63b83d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
