--- Assets/Scenes/test3.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 43b299beb90e89349a8d1375ae49d629
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
