--- Assets/Scenes/test3/OcclusionCullingData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 78f9f41ce734eb240ac1a191dd909c40
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 36300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
